requirements/base.txt
Django==2.2
pytz==2019.1
sqlparse==0.3.0
django-model-utils==3.1.2
djangorestframework==3.9.2
Markdown==3.1
python-dateutil==2.8.0
six==1.12.0
text-unidecode==1.2
MarkupSafe==1.1.1
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0
idna==2.8
itypes==1.1.0
jinja2==2.10.1
openapi-codec==1.3.2
requests==2.21.0
simplejson==3.16.0
uritemplate==3.0.0
urllib3==1.24.1